Updated React to version 2.0.3 and added some more mp3 code but it can't play yet
