+++ package.json
@@ -6,11 +6,12 @@
     "axios": "^0.18.0",
     "bootstrap": "^4.1.3",
     "merge": "^1.2.1",
+    "node-sass": "^4.10.0",
     "npm": "^6.4.1",
     "react": "^16.5.2",
     "react-dom": "^16.5.2",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "1.1.4",
+    "react-scripts": "2.0.3",
     "reactstrap": "^6.4.0"
   },
   "scripts": {
@@ -18,5 +19,11 @@
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
-  }
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
