+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "Cardsite",
+  "name": "Tim's webpage for cards",
   "icons": [
     {
       "src": "favicon.ico",
