+++ src/Content/MusicPlayer/NewStyles.scss
@@ -1,5 +1,3 @@
-@import "https://fonts.googleapis.com/css?family=Roboto:300,400,500,700";
-
 // Colors
 $black: #121212;
 $white: #ffffff;
@@ -42,11 +40,11 @@ $blue: #1d2d3a;
 body {
   align-items: center;
   background: darken($blue, 10);
-  display: flex;
+  //display: flex;
   font-family: "Roboto", sans-serif;
-  color: white;
-  height: 100vh;
-  justify-content: center;
+  //color: white;
+  min-height: 100vh;
+  //justify-content: center;
 }
 
 .Player {
@@ -57,6 +55,7 @@ body {
   height: 667px;
   position: relative;
   width: 375px;
+  margin-bottom: 3em;
 
   .Background {
     width: 150%;
@@ -78,6 +77,7 @@ body {
     text-transform: uppercase;
     letter-spacing: 0.1em;
     position: relative;
+    color: $white;
   }
 
   .Artwork {
@@ -102,17 +102,20 @@ body {
       font-size: 24px;
       margin-bottom: 10px;
       font-weight: 300;
+      color: $white;
     }
 
     .Artist {
       font-size: 16px;
       margin-bottom: 5px;
       font-weight: 500;
+      color: $white;
     }
 
     .Album {
       font-size: 12px;
       opacity: 0.5;
+      color: $white;
     }
   }
 
@@ -152,6 +155,7 @@ body {
     pointer-events: none;
     z-index: 2;
     width: 100%;
+    color: $white;
 
     .Time {
       font-size: 12px;
